+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Detect infinite loops in chain configuration templates
 - Duplicated chains in a recipe will only be rendered once [#118](https://github.com/LucasPickering/slumber/issues/118)
 - Never trigger chained requests when rendering template previews in the TUI
+- Use a different log file for each session [#61](https://github.com/LucasPickering/slumber/issues/61)
 
 ## [1.6.0] - 2024-07-07